move `load` handler to its own module
