+++ unison-cli/src/Unison/Codebase/Editor/Input.hs
@@ -52,6 +52,7 @@ import Unison.Util.Pretty qualified as P
 data Event
   = UnisonFileChanged SourceName Source
   | IncomingRootBranch (Set CausalHash)
+  deriving stock (Show)
 
 type Source = Text -- "id x = x\nconst a b = a"
